Move `Depth` to utilities
